Revert "feat(account): user account page redesign (#206)"
This reverts commit ff91a59246ee6fb8894a9d43c8bd679ebff19518.
