--- app/(app)/account/_components/profile/UsernameButton.tsx
@@ -1,135 +1,0 @@
-'use client';
-
-import { zodResolver } from '@hookform/resolvers/zod';
-import { Loader2, Save } from 'lucide-react';
-import { useEffect, useState } from 'react';
-import { useForm } from 'react-hook-form';
-import { z } from 'zod';
-
-import { Button } from '@/components/ui/button';
-import ButtonItem from '@/components/ui/Button/ButtonList/ButtonItem';
-import { Dialog, DialogFooter, DialogHeader, DialogContent, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
-import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
-import { Input } from '@/components/ui/input';
-import useAuth from '@/context/auth/useAuth';
-import useCsrf from '@/context/csrf/useCsrf';
-import useErrorToast from '@/hooks/error/useErrorToast';
-import { updateAccount } from '@/services/auth.service';
-import { ApiError } from '@/utils/api/error';
-
-
-const UsernameFormSchema = z.object({ username: z.string().min(1, 'Required.') });
-
-type UsernameFormValues = z.infer<typeof UsernameFormSchema>;
-
-export const UsernameButton = () => {
-
-	const [ isLoading, setIsLoading ] = useState(false);
-	const [ isDialogOpen, setIsDialogOpen ] = useState(false);
-
-	const { currentUser, updateCurrentUser } = useAuth();
-	const { csrfToken } = useCsrf();
-	const { triggerErrorToast } = useErrorToast();
-
-	const form = useForm<UsernameFormValues>({
-		resolver: zodResolver(UsernameFormSchema),
-		defaultValues: { username: currentUser?.username || '' },
-		mode: 'onTouched',
-	});
-
-	useEffect(() => {
-		if (currentUser) {
-			form.setValue('username', currentUser.username || '');
-		}
-	// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [ currentUser?.id ]);
-
-	const handleOpenDialog = () => setIsDialogOpen(true);
-
-	const handleSubmit = async (values: UsernameFormValues) => {
-		const { username } = values;
-		if (!csrfToken) {
-			triggerErrorToast({
-				title: 'Error',
-				message: 'Invalid CSRF token.',
-			});
-			return;
-		}
-		try {
-			setIsLoading(true);
-			const updatedUser = await updateAccount({ username }, { csrfToken });
-			await updateCurrentUser(updatedUser);
-			form.reset({ username: updatedUser.username || '' });
-			setIsDialogOpen(false);
-		} catch (error) {
-			const apiError = error as ApiError<unknown>;
-			triggerErrorToast(apiError, form);
-		} finally {
-			setIsLoading(false);
-		}
-	};
-
-	const handleDialogOpenChange = (open: boolean) => {
-		if (!isLoading) {
-			setIsDialogOpen(open);
-		}
-	};
-
-	return (
-		<>
-			<Dialog
-				open={ isDialogOpen }
-				onOpenChange={ handleDialogOpenChange }
-			>
-				<DialogTrigger asChild>
-					<ButtonItem
-						value={ currentUser?.username }
-						onClick={ handleOpenDialog }
-					>
-						Username
-					</ButtonItem>
-				</DialogTrigger>
-				<DialogContent className="sm:max-w-[425px]">
-					<Form { ...form }>
-						<form onSubmit={ form.handleSubmit(handleSubmit) }>
-							<DialogHeader>
-								<DialogTitle>Username</DialogTitle>
-							</DialogHeader>
-							<div className="py-4">
-								<FormField
-									control={ form.control }
-									name="username"
-									render={ ({ field }) => (
-										<FormItem>
-											<FormLabel>Username</FormLabel>
-											<FormControl>
-												<Input
-													placeholder="John Doe"
-													type="text"
-													{ ...field }
-												/>
-											</FormControl>
-											<FormMessage />
-										</FormItem>
-									) }
-								/>
-							</div>
-							<DialogFooter>
-								<Button
-									className="gap-2"
-									disabled={ isLoading }
-									type="submit"
-								>
-									{ isLoading ? <Loader2 className="h-4 w-4 animate-spin" /> : <Save size="16" /> }
-									Save
-								</Button>
-							</DialogFooter>
-						</form>
-					</Form>
-				</DialogContent>
-			</Dialog>
-		</>
-	);
-};
-
-export default UsernameButton;
