+++ next.config.mjs
@@ -4,7 +4,6 @@ const withBundleAnalyzer = NextBundleAnalyzer({ enabled: process.env.ANALYZE ===
   
 export default withBundleAnalyzer({
 	eslint: { ignoreDuringBuilds: true },
-	// typescript: { ignoreBuildErrors: true },
 	images: {
 		remotePatterns: [
 			{
