+++ package.json
@@ -87,7 +87,6 @@
 		"@radix-ui/react-dropdown-menu": "^2.0.5",
 		"@radix-ui/react-label": "^2.0.2",
 		"@radix-ui/react-popover": "^1.0.6",
-		"@radix-ui/react-radio-group": "^1.1.3",
 		"@radix-ui/react-select": "^1.2.2",
 		"@radix-ui/react-separator": "^1.0.3",
 		"@radix-ui/react-slot": "^1.0.2",
@@ -112,7 +111,6 @@
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.0.0",
 		"cmdk": "^0.2.0",
-		"date-fns": "^2.30.0",
 		"edge-csrf": "^1.0.4",
 		"i": "^0.3.7",
 		"jsonwebtoken": "^9.0.1",
@@ -126,7 +124,6 @@
 		"nodemailer": "^6.9.7",
 		"npm": "^9.8.1",
 		"react": "^18.2.0",
-		"react-day-picker": "^8.9.1",
 		"react-dom": "^18.2.0",
 		"react-email": "1.9.4",
 		"react-hook-form": "^7.47.0",
